Move the utils into another file.
