+++ inputs/__init__.py
@@ -2433,33 +2433,3 @@ class MicroBitListener(BaseListener):
         self.update_timeval()
         self.events.append(self.sync_marker(self.timeval))
         self.write_to_pipe(self.events)
-
-
-devices = DeviceManager()  # pylint: disable=invalid-name
-
-
-def get_key():
-    """Get a single keypress from a keyboard."""
-    try:
-        keyboard = devices.keyboards[0]
-    except IndexError:
-        raise UnpluggedError("No keyboard found.")
-    return keyboard.read()
-
-
-def get_mouse():
-    """Get a single movement or click from a mouse."""
-    try:
-        mouse = devices.mice[0]
-    except IndexError:
-        raise UnpluggedError("No mice found.")
-    return mouse.read()
-
-
-def get_gamepad():
-    """Get a single action from a gamepad."""
-    try:
-        gamepad = devices.gamepads[0]
-    except IndexError:
-        raise UnpluggedError("No gamepad found.")
-    return gamepad.read()
